KOGITO-6168/6194 : Additional JS methods

+++ lienzo-core/src/main/java/com/ait/lienzo/client/core/types/JsCanvas.java
@@ -276,8 +276,8 @@ public class JsCanvas implements JsCanvasNodeLister {
 
         if (absoluteLocation != null) {
 
-            int width = layer.getViewport().getWidth();
-            int height = layer.getViewport().getHeight();
+            double width = layer.getViewport().getWidth();
+            double height = layer.getViewport().getHeight();
 
             NFastArrayList<Double> dimensions = getDimensions(UUID);
             double nodeWidth = 0;
